[Java] emit right fn signature for handler of no arguments
Anonymous event handlers need not consume the event, such as here:

```
     1  type t = (a : seq[int]);
     2  event ev : t;
     3
     4  event ev2 : int;
     5
     6  spec foo observes ev, ev2 {
     7    start state Init {
     8      on ev, ev2 do {
     9
    10      }
    11    }
    12
    13  }
```

Even though `ev` and `ev2` have incompatable payload types, this code is still
safe since we never do anything with it.  However, the current code generator
would incorrectly emit a parameter to the generated method, which would cause
the generated Java code to fail typechecking, like so.

```
 58         private void Anon(ev pEvent) {
 59
 60         }
 ...
 64             addState(new State.Builder(INIT_STATE)
 65                 .isInitialState(true)
 66                 .withEvent(ev.class, this::Anon)
 67                 .withEvent(ev2.class, this::Anon) // <-- This fails to typecheck!
 68                 .build());
```
This patch addresses cases where the Function signature's parameter list is
empty, and does not emit an argument for it.  At monitor construction time, the
event handler for `ev` and `ev2` both consume the event, as is required for
typechecking to succeed, but discards the result before calling the actual
handler.

```
 58         private void Anon() {
 59
 60         }
 ...
 64             addState(new State.Builder(INIT_STATE)
 65                 .isInitialState(true)
 66                 .withEvent(ev.class, __ -> Anon(); )
 67                 .withEvent(ev2.class, __ -> Anon(); )
 68                 .build());
```