[PCSharpRuntime] Log escaped string representations of PrtValues
This is an attempt to address https://github.com/p-org/P/issues/447 , but we
should think carefully about whether we've covered our bases on this.

Previously, the PLogFormatter would emit the direct string representation
of any PrtValues in an event payload.  This means that there could be ambiguity
owing to PrtStrings not being escaped.  For instance, the P statement

```
send f, ev, (s = "a:string, with:\"problems",);
```

Would yield a log line of the form:

```
<SendLog> 'PImplementation.Driver(2)' in state 'Init_1' sent event 'ev with payload (<s:a:string, with:"problems, >)' to 'Foo(3)'
```

Which is ambiguous in terms of what the NamedTuple arity is, and also
results in an unclosed string literal.

This patch introduces a variation on ToString(), called ToEscapedString(),
which ensures that problematic metacharacters are quoted, and PrtStrings
are themselves enclosed in double-quotes.  The above log line would now be

```
<SendLog> 'PImplementation.Driver(2)' in state 'Init_1' sent event 'ev with payload (<s:"a:string, with:\"problems", >)' to 'Foo(3)'.
```

This change could not be done as a modification of ToString(), since the P
runtime relies on a "naked" representation of the value.