[Java] Uniquify the top-level class name
We use the top-level class as a sort of namespace to contain all the
nested classes for a file.  If a P programmer gives a Monitor the
same name as the P file name, Java will complain about a duplicate
class.

(We cannot change the top level class without also changing the
filename that is emitted)