[Java] FFI: stub out checked exceptions in foreign function stubs
If a foreign function wants to change state or raise an event, it
does so by throwing a checked exception.  However, this means that
the foreign function _itself_ needs to be marked as throwing a
checked exception.

We can't generate these marks like we do with the P -> Java extraction
code because of course we don't know what the developer will put in
there.  And, we can't be conservative and "taint" all foreign functions
as "potentially" transitioning or throwing because that invalidates
type-level constraints in the PRT (i.e. that you can't transition
in an exit handler).

So, just put them there but commented out, so developers know what
to do if they do indeed want to transition or raise an event.