[CLI] Remove --target/--generate and add --projname/--mode
Updates CLI options to remove --target and --generate options. Instead, --target is renamed to --projname and --mode is used to configure the compiler backend (default: bugfinding)

Adds support for searching for *.dll in current directory with p check command when path is not given

Adds fallback support for searching for adding local P files in the current directory with p compile command when no P file is given through *.pproj and --pfiles options.