[PCompiler] Always rewrite .csproj (or pom.xml)
The .csproj file generated by P compiler for C# backend hardcodes the PCSharpRuntime and PCheckerCore versions based on the compiler version. The .csproj file is not overwritten if already present. So if anyone upgrades the P version, they have to manually delete the compiler output folder. Otherwise, they get an error on running the p check command. With this change, we force build configs to be overwritten always.